Make join depend completely on tracker
This is temporary until there's a solution for a DHT
based discovery. This prevents peers from answering
join requests for networks they don't own or know
which lead to strange behaviour.

A root peer has to respond in a given timeout (2 seconds).
It's currently not possible to change this timeout.